updated terms for catalog
