+++ apps/catalog/catalog/src/app/dashboard/dashboard.module.ts
@@ -80,19 +80,19 @@ const routes: Routes = [
       },
       {
         path: 'about',
-        loadChildren: () => import('./pages/about-page/about.module').then(m => m.AboutModule)
+        loadChildren: () => import('@blockframes/ui/static-informations/about/about.module').then(m => m.AboutModule)
       },
       {
         path: 'who-are-we',
-        loadChildren: () => import('./pages/team-page/team.module').then(m => m.TeamModule)
+        loadChildren: () => import('@blockframes/ui/static-informations/team/team.module').then(m => m.TeamModule)
       },
       {
         path: 'contact',
-        loadChildren: () => import('./pages/contact-page/contact.module').then(m => m.ContactModule)
+        loadChildren: () => import('@blockframes/ui/static-informations/contact/contact.module').then(m => m.ContactModule)
       },
       {
         path: 'terms',
-        loadChildren: () => import('./pages/privacy-page/privacy.module').then(m => m.PrivacyModule)
+        loadChildren: () => import('@blockframes/ui/static-informations/privacy/privacy.module').then(m => m.PrivacyModule)
       },
     ]
   },
