--- apps/catalog/catalog/src/app/dashboard/pages/about-page/about-page.component.html
@@ -1,48 +1,0 @@
-<header>
-  <h1>About</h1>
-</header>
-
-<section fxLayout="row" fxLayoutAlign="center">
-  <mat-card>
-    <mat-card-content>
-      <article>
-        <h3>What is Archipel Content ?</h3>
-        <p>
-          Archipel Content is the first marketplace offering package deals for TV and Film rights.
-          The Archipel Content platform gives Content Buyers access to a large consolidated library, composed of several Sales Agents’ titles.
-          That way, we offer Content Buyers a way to purchase more titles in a more efficient way, and Sales Agents the chance to create new opportunities by mutualizing their assets.
-          Our goal is to streamline the whole licensing process: by offering an efficient tool to trade rights, while instantaneously redistributing revenues to the rightful stakeholders, thanks to Blockframes, a blockchain-powered value chain.
-        </p>
-      </article>
-
-      <article>
-        <h3>What is Blockframes?</h3>
-        <p>
-          Blockframes is a technological ecosystem for the film industry, aiming to bring transparency and efficiency to independent cinema.
-          Concretely, Blockframes replicates a film’s value chain on the blockchain, to streamline all transactions linked to the film’s lifecycle and waterfall.
-        </p>
-      </article>
-
-      <article>
-        <h3>Sounds nice... but what is the blockchain exactly?</h3>
-        <p>
-          “The blockchain is an incorruptible digital ledger of economic transactions that can be programmed to record not just financial transactions, but virtually everything of value.” – Blockchain Revolution (2016) by Don & Alex Tapscott.
-          Basically, it’s a time-stamped, immutable, transparent, secured and shared database.
-          What’s so special about it is that it brings decentralization, transparency, traceability, security, consensus, trust, automation, efficiency… Concepts that we could undeniably use in the cinema industry!
-        </p>
-      </article>
-
-      <article>
-        <h3>Where does this project come from? </h3>
-        <p>
-          The Blockframes project comes directly from the experience and expertise of our mother company, Logical Pictures.
-          Logical Pictures is an innovating film production and financing company, launched in 2016 by serial entrepreneur Frédéric Fiore and his partner Yannick Bossenmeyer.
-          As they were still new to the cinema industry, they identified quite a few bottlenecks at different levels of the value chain. With a non-formatted and open mindset, they started looking for solutions to answers these inefficiencies.
-          They got particularly curious about the blockchain technology and quickly realized all the opportunities that it could offer to the cinema industry.
-          They developed the idea of Blockframes and launched a successful proof of concept on the animation film Terra Willy in 2018.
-          In 2019, they launched Cascade8, the start-up company that’s now in charge of developing these blockchain-powered tech solutions for the cinema industry. With Blockframes as a background, we’re developing a suite of DApps (“Decentralized Apps”: basically, blockchain-powered apps) that interact with each other. Archipel Content is one of them.
-        </p>
-      </article>
-    </mat-card-content>
-  </mat-card>
-</section>
