--- apps/catalog/catalog/src/app/dashboard/pages/about-page/about-page.component.ts
@@ -1,10 +1,0 @@
-import { ChangeDetectionStrategy, Component } from "@angular/core";
-
-@Component({
-  selector: 'catalog-about-page',
-  templateUrl: './about-page.component.html',
-  styleUrls: ['./about-page.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush
-})
-
-export class AboutPageComponent {}
