--- apps/catalog/catalog/src/app/dashboard/pages/about-page/about.module.ts
@@ -1,27 +1,0 @@
-// Angular
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { RouterModule } from '@angular/router';
-
-// Pages
-import { AboutPageComponent } from './about-page.component';
-
-// Material
-import { MatCardModule } from '@angular/material/card';
-import { FlexLayoutModule } from '@angular/flex-layout';
-
-@NgModule({
-  declarations: [AboutPageComponent],
-  imports: [
-    CommonModule,
-    MatCardModule,
-    FlexLayoutModule,
-    RouterModule.forChild([
-      {
-        path: '',
-        component: AboutPageComponent
-      }
-    ])
-  ]
-})
-export class AboutModule {}
