--- apps/catalog/catalog/src/app/dashboard/pages/contact-page/contact-page.component.scss
@@ -1,50 +1,0 @@
-@import '~@angular/flex-layout/mq';
-
-section {
-  min-height: calc(100vh - (var(--toolbarHeight) * 2) - 30px); // TODO #1212
-}
-
-header {
-  margin-left: 24px;
-}
-
-h1 {
-  margin-top: 24px;
-  margin-bottom: 8px;
-}
-
-h3 {
-  margin-bottom: 4px;
-}
-
-p {
-  margin-top: 8px;
-  margin-bottom: 34px;
-  &.text-margin {
-    margin-top: 150px;
-  }
-}
-
-article {
-  width: 50%;
-}
-
-mat-form-field {
-  margin-left: 24px;
-  margin-right: 24px;
-}
-
-@include layout-bp(lt-lg) {
-  mat-card {
-    width: 80%;
-    margin-top: 50px;
-  }
-}
-
-@include layout-bp(xs) {
-  mat-card {
-    width: 100%;
-    margin-top: 30px;
-    margin-bottom: 30px;
-  }
-}
