--- apps/catalog/catalog/src/app/dashboard/pages/contact-page/contact.module.ts
@@ -1,39 +1,0 @@
-// Angular
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { RouterModule } from '@angular/router';
-import { ReactiveFormsModule } from '@angular/forms';
-import { GoogleMapsModule } from '@angular/google-maps';
-
-// Pages
-import { ContactPageComponent } from './contact-page.component';
-
-// Material
-import { MatCardModule } from '@angular/material/card';
-import { FlexLayoutModule } from '@angular/flex-layout';
-import { MatFormFieldModule } from '@angular/material/form-field';
-import { MatInputModule } from '@angular/material/input';
-import { MatButtonModule } from '@angular/material/button';
-import { MatIconModule } from '@angular/material/icon';
-
-@NgModule({
-  declarations: [ContactPageComponent],
-  imports: [
-    CommonModule,
-    MatCardModule,
-    FlexLayoutModule,
-    ReactiveFormsModule,
-    MatFormFieldModule,
-    MatIconModule,
-    MatButtonModule,
-    MatInputModule,
-    GoogleMapsModule,
-    RouterModule.forChild([
-      {
-        path: '',
-        component: ContactPageComponent
-      }
-    ])
-  ]
-})
-export class ContactModule {}
