--- apps/catalog/catalog/src/app/dashboard/pages/privacy-page/privacy-page.component.ts
@@ -1,12 +1,0 @@
-import { ChangeDetectionStrategy, Component } from "@angular/core";
-
-@Component({
-  selector: 'catalog-privacy-page',
-  template: '<auth-terms-conditions></auth-terms-conditions>',
-  changeDetection: ChangeDetectionStrategy.OnPush
-})
-
-export class PrivacyPageComponent {
-
-  constructor() {}
-}
