--- apps/catalog/catalog/src/app/dashboard/pages/team-page/team-page.component.scss
@@ -1,70 +1,0 @@
-@import '~@angular/flex-layout/mq';
-
-header {
-  padding: 100px 0;
-  // TODO: replace image when the real one is available.
-  background-image: url('/assets/images/catalog-marketplace-banner.png');
-  background-position: center center;
-  background-attachment: fixed;
-  background-size: cover;
-}
-
-mat-card {
-  margin-top: 50px;
-  width: 1000px;
-  margin-bottom: 50px;
-
-  .partner-logo {
-    height: 40px;
-    margin-right: 40px;
-  }
-
-  h4 {
-    margin-bottom: 0;
-  }
-
-  .margin-top {
-    margin-top: 15px;
-  }
-
-  .photo-contact {
-    height: 170px;
-    width: 170px;
-    margin-right: 20px;
-    border-radius: 100%;
-  }
-}
-
-@include layout-bp(xs) {
-  mat-card {
-    width: 100%;
-    margin-top: 30px;
-    margin-bottom: 30px;
-  }
-
-  .partner-logo {
-    max-width: 200px;
-    margin-right: 0;
-    margin-bottom: 25px;
-  }
-}
-
-@include layout-bp(sm) {
-  mat-card {
-    width: 80%;
-    margin-top: 30px;
-  }
-
-  .partner-logo {
-    max-width: 200px;
-    margin-right: 0;
-    margin-bottom: 25px;
-  }
-}
-
-@include layout-bp(md) {
-  mat-card {
-    width: 80%;
-    margin-top: 30px;
-  }
-}
