--- apps/catalog/catalog/src/app/dashboard/pages/team-page/team-page.component.ts
@@ -1,10 +1,0 @@
-import { Component, ChangeDetectionStrategy  } from "@angular/core";
-
-@Component({
-  selector: 'catalog-team-page',
-  templateUrl: './team-page.component.html',
-  styleUrls: ['./team-page.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush
-})
-
-export class TeamPageComponent {}
