--- apps/catalog/catalog/src/app/dashboard/pages/team-page/team.module.ts
@@ -1,27 +1,0 @@
-// Angular
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { RouterModule } from '@angular/router';
-
-// Pages
-import { TeamPageComponent } from './team-page.component';
-
-// Material
-import { MatCardModule } from '@angular/material/card';
-import { FlexLayoutModule } from '@angular/flex-layout';
-
-@NgModule({
-  declarations: [TeamPageComponent],
-  imports: [
-    CommonModule,
-    MatCardModule,
-    FlexLayoutModule,
-    RouterModule.forChild([
-      {
-        path: '',
-        component: TeamPageComponent
-      }
-    ])
-  ]
-})
-export class TeamModule {}
