--- apps/catalog/catalog/src/app/marketplace/landing/about/about.component.scss
@@ -1,31 +1,0 @@
-@import '~@angular/flex-layout/mq';
-
-header {
-  padding: 100px 0;
-  background-image: url('/assets/images/catalog-marketplace-banner.png');
-  background-position: center center;
-  background-attachment: fixed;
-  background-size: cover;
-  text-align: center;
-}
-
-mat-card {
-  margin-top: 50px;
-  margin-bottom: 50px;
-  width: 1000px;
-}
-
-@include layout-bp(lt-lg) {
-  mat-card {
-    width: 80%;
-    margin-top: 50px;
-  }
-}
-
-@include layout-bp(xs) {
-  mat-card {
-    width: 100%;
-    margin-top: 30px;
-    margin-bottom: 30px;
-  }
-}
