--- apps/catalog/catalog/src/app/marketplace/landing/about/about.component.ts
@@ -1,10 +1,0 @@
-import { ChangeDetectionStrategy, Component } from "@angular/core";
-
-@Component({
-  selector: 'catalog-about',
-  templateUrl: './about.component.html',
-  styleUrls: ['./about.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush
-})
-
-export class AboutComponent {}
