--- apps/catalog/catalog/src/app/marketplace/landing/about/about.module.ts
@@ -1,27 +1,0 @@
-// Angular
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { RouterModule } from '@angular/router';
-
-// Pages
-import { AboutComponent } from './about.component';
-
-// Material
-import { MatCardModule } from '@angular/material/card';
-import { FlexLayoutModule } from '@angular/flex-layout';
-
-@NgModule({
-  declarations: [AboutComponent],
-  imports: [
-    CommonModule,
-    MatCardModule,
-    FlexLayoutModule,
-    RouterModule.forChild([
-      {
-        path: '',
-        component: AboutComponent
-      }
-    ])
-  ]
-})
-export class AboutModule {}
