--- apps/catalog/catalog/src/app/marketplace/landing/contact/contact.component.html
@@ -1,27 +1,0 @@
-<section>
-  <header>
-    <h1>Contact</h1>
-  </header>
-  <article fxLayout="row" fxLayoutAlign="center">
-    <mat-card>
-      <mat-card-content>
-        <article>
-          <h3>Contact</h3>
-          <p>
-            The Archipel Content platform is currently animated by Marie. You can contact her if you have any questions
-            regarding Archipel Content. She’ll be happy to answer you and help you with anything you need!
-          </p>
-        </article>
-
-        <article fxLayout="row" fxLayout.lt-sm="column">
-          <img class="photo-contact" src="/assets/photo/marie-photo-avatar.jpg">
-          <div fxLayout="column">
-            <h4>Marie</h4>
-            <span>Altmayer</span>
-            <span class="margin-top"><span class="bf-dark-text">Email address</span> maltmayer@cascade8.com</span>
-          </div>
-        </article>
-      </mat-card-content>
-    </mat-card>
-  </article>
-</section>
