--- apps/catalog/catalog/src/app/marketplace/landing/contact/contact.component.scss
@@ -1,50 +1,0 @@
-@import '~@angular/flex-layout/mq';
-
-section {
-  min-height: calc(100vh - (var(--toolbarHeight) * 2) - 30px); // TODO #1212
-}
-
-header {
-  padding: 100px 0;
-  background-image: url('/assets/images/catalog-marketplace-banner.png');
-  background-position: center center;
-  background-attachment: fixed;
-  background-size: cover;
-  text-align: center;
-}
-
-mat-card {
-  margin-top: 50px;
-  width: 1000px;
-  margin-bottom: 50px;
-
-  h4 {
-    margin-bottom: 0;
-  }
-
-  .margin-top {
-    margin-top: 15px;
-  }
-
-  .photo-contact {
-    height: 170px;
-    width: 170px;
-    margin-right: 20px;
-    border-radius: 100%;
-  }
-}
-
-@include layout-bp(lt-lg) {
-  mat-card {
-    width: 80%;
-    margin-top: 50px;
-  }
-}
-
-@include layout-bp(xs) {
-  mat-card {
-    width: 100%;
-    margin-top: 30px;
-    margin-bottom: 30px;
-  }
-}
