--- apps/catalog/catalog/src/app/marketplace/landing/contact/contact.component.ts
@@ -1,10 +1,0 @@
-import { ChangeDetectionStrategy, Component } from "@angular/core";
-
-@Component({
-  selector: 'catalog-contact',
-  templateUrl: './contact.component.html',
-  styleUrls: ['./contact.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush
-})
-
-export class ContactComponent {}
