--- apps/catalog/catalog/src/app/marketplace/landing/contact/contact.module.ts
@@ -1,27 +1,0 @@
-// Angular
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { RouterModule } from '@angular/router';
-
-// Pages
-import { ContactComponent } from './contact.component';
-
-// Material
-import { MatCardModule } from '@angular/material/card';
-import { FlexLayoutModule } from '@angular/flex-layout';
-
-@NgModule({
-  declarations: [ContactComponent],
-  imports: [
-    CommonModule,
-    MatCardModule,
-    FlexLayoutModule,
-    RouterModule.forChild([
-      {
-        path: '',
-        component: ContactComponent
-      }
-    ])
-  ]
-})
-export class ContactModule {}
