--- apps/catalog/catalog/src/app/marketplace/landing/privacy/privacy.component.ts
@@ -1,12 +1,0 @@
-import { ChangeDetectionStrategy, Component } from "@angular/core";
-
-@Component({
-  selector: 'catalog-privacy',
-  template: '<auth-terms-conditions></auth-terms-conditions>',
-  changeDetection: ChangeDetectionStrategy.OnPush
-})
-
-export class PrivacyComponent {
-
-  constructor() {}
-}
