--- apps/catalog/catalog/src/app/marketplace/landing/team/team.component.ts
@@ -1,10 +1,0 @@
-import { Component, ChangeDetectionStrategy  } from "@angular/core";
-
-@Component({
-  selector: 'catalog-team',
-  templateUrl: './team.component.html',
-  styleUrls: ['./team.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush
-})
-
-export class TeamComponent {}
