--- apps/catalog/catalog/src/app/marketplace/landing/team/team.module.ts
@@ -1,27 +1,0 @@
-// Angular
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { RouterModule } from '@angular/router';
-
-// Pages
-import { TeamComponent } from './team.component';
-
-// Material
-import { MatCardModule } from '@angular/material/card';
-import { FlexLayoutModule } from '@angular/flex-layout';
-
-@NgModule({
-  declarations: [TeamComponent],
-  imports: [
-    CommonModule,
-    MatCardModule,
-    FlexLayoutModule,
-    RouterModule.forChild([
-      {
-        path: '',
-        component: TeamComponent
-      }
-    ])
-  ]
-})
-export class TeamModule {}
