+++ apps/festival/festival/src/index.html
@@ -4,6 +4,7 @@
     <meta charset="utf-8" />
     <title>Archipel Market</title>
     <base href="/" />
+    <script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyDuWt_jQha3JubeH5oBkRzn-0X1ggi2DVk"></script>
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <link rel="icon" type="image/x-icon" href="favicon.ico" />
     <meta name="theme-color" content="#1976d2">
