+++ libs/ui/src/lib/layout/footer/footer.component.ts
@@ -1,5 +1,5 @@
 // Angular
-import { Component, ChangeDetectionStrategy, ChangeDetectorRef, OnInit } from '@angular/core';
+import { Component, ChangeDetectionStrategy, OnInit } from '@angular/core';
 
 // Blockframes
 import { getAppLocation } from '@blockframes/utils/helpers';
@@ -18,9 +18,12 @@ import { map } from 'rxjs/operators';
 export class FooterComponent implements OnInit {
   public section$: Observable<'dashboard' | 'marketplace'>;
 
-  constructor(private routerQuery: RouterQuery, private cdr: ChangeDetectorRef) { }
+  public isFestival = false;
+
+  constructor(private routerQuery: RouterQuery) { }
 
   ngOnInit() {
     this.section$ = this.routerQuery.select('state').pipe(map(data => getAppLocation(data.url)));
+    this.isFestival = this.routerQuery.getData<string>('app') === 'festival' ? true : false;
   }
 }
