+++ libs/ui/src/lib/static-informations/contact/contact.component.scss
@@ -5,33 +5,33 @@ section {
 }
 
 header {
-  padding: 100px 0;
-  background-image: url('/assets/images/catalog-marketplace-banner.png');
-  background-position: center center;
-  background-attachment: fixed;
-  background-size: cover;
-  text-align: center;
+  margin-left: 24px;
 }
 
-mat-card {
-  margin-top: 50px;
-  width: 1000px;
-  margin-bottom: 50px;
+h1 {
+  margin-top: 24px;
+  margin-bottom: 8px;
+}
 
-  h4 {
-    margin-bottom: 0;
-  }
+h3 {
+  margin-bottom: 4px;
+}
 
-  .margin-top {
-    margin-top: 15px;
+p {
+  margin-top: 8px;
+  margin-bottom: 34px;
+  &.text-margin {
+    margin-top: 150px;
   }
+}
 
-  .photo-contact {
-    height: 170px;
-    width: 170px;
-    margin-right: 20px;
-    border-radius: 100%;
-  }
+article {
+  width: 50%;
+}
+
+mat-form-field {
+  margin-left: 24px;
+  margin-right: 24px;
 }
 
 @include layout-bp(lt-lg) {
