--- libs/ui/src/lib/static-informations/team-page/team-page.component.html
@@ -1,26 +1,0 @@
-<header fxLayout="column" fxLayoutAlign="center center">
-  <h1>Who are we?</h1>
-</header>
-
-<section fxLayout="row" fxLayoutAlign="center">
-  <mat-card>
-    <mat-card-content>
-      <article>
-        <h3>The team</h3>
-        <p>
-            The team behind Archipel Content is the Cascade8 team, composed of 15 people with backgrounds in tech development, design, cinema and business... working from Paris, Lyon and Berlin!
-        </p>
-      </article>
-
-      <article>
-        <h3>Our partners</h3>
-        <div fxLayout="row" fxLayout.lt-md="column">
-          <img class="partner-logo" src="/assets/logo/pulsar.png">
-          <img class="partner-logo" src="/assets/logo/bpi.png">
-          <img class="partner-logo" src="/assets/logo/creativemedia.png">
-          <img class="partner-logo" src="/assets/logo/cnc.png">
-        </div>
-      </article>
-    </mat-card-content>
-  </mat-card>
-</section>
