+++ libs/ui/src/lib/static-informations/team/team.component.ts
@@ -1,7 +1,7 @@
 import { Component, ChangeDetectionStrategy  } from "@angular/core";
 
 @Component({
-  selector: 'catalog-team',
+  selector: 'bf-team',
   templateUrl: './team.component.html',
   styleUrls: ['./team.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush
