+++ libs/ui/src/lib/static-informations/terms/terms.component.ts
@@ -1,0 +1,12 @@
+import { Component, OnInit, ChangeDetectionStrategy } from '@angular/core';
+
+@Component({
+  selector: 'bf-terms',
+  template: `<auth-terms-conditions></auth-terms-conditions>`,
+  changeDetection: ChangeDetectionStrategy.OnPush
+})
+export class TermsComponent implements OnInit {
+  constructor() { }
+
+  ngOnInit() { }
+}
