+++ libs/user/src/lib/auth/components/terms-conditions/terms-conditions.component.ts
@@ -1,4 +1,5 @@
 import { ChangeDetectionStrategy, Component } from "@angular/core";
+import { RouterQuery } from '@datorama/akita-ng-router-store';
 
 @Component({
   selector: 'auth-terms-conditions',
@@ -6,5 +7,9 @@ import { ChangeDetectionStrategy, Component } from "@angular/core";
   styleUrls: ['./terms-conditions.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-
-export class TermsConditionsComponent {}
+export class TermsConditionsComponent {
+  public isFestival: boolean;
+  constructor(private routerQuery: RouterQuery) {
+    this.isFestival = this.routerQuery.getData<string>('app') === 'festival' ? true : false;
+  }
+}
